requirements.txt
Pillow==8.0.1
PyYAML==5.3.1
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
iso8601==0.1.13
kiwisolver==1.3.0
matplotlib==3.3.2
numpy==1.19.3
pip==20.2.4
pyparsing==2.4.7
python-dateutil==2.8.1
serial==0.0.97
six==1.15.0